Stop/start background fetch
